Use size_t as operator[] argument type
This better mimics std::vector::operator[] and
fixes a warning with MSVC 64bit.

No functional change.

+++ src/thread.h
@@ -108,7 +108,7 @@ public:
   void init(); // No c'tor, Threads object is global and engine shall be fully initialized
   ~ThreadPool();
 
-  Thread& operator[](int id) { return *threads[id]; }
+  Thread& operator[](size_t id) { return *threads[id]; }
   bool use_sleeping_threads() const { return useSleepingThreads; }
   int min_split_depth() const { return minimumSplitDepth; }
   size_t size() const { return threads.size(); }
